Updated shared appsettings
