+++ src/backend/Aspire/Board.ServiceDefaults/Board.ServiceDefaults.csproj
@@ -18,6 +18,7 @@
         <PackageReference Include="Aspire.Microsoft.EntityFrameworkCore.SqlServer" />
         <PackageReference Include="Serilog.Settings.Configuration" />
         <PackageReference Include="Serilog.Extensions.Logging" />
+        <PackageReference Include="Serilog.Sinks.Console" />
     </ItemGroup>
 
 	<ItemGroup>
