+++ src/backend/Aspire/Board.ServiceDefaults/appsettings.shared.json
@@ -1,5 +1,33 @@
 {
   "ConnectionStrings": {
     "BoardDbConnectionString": "Data Source=127.0.0.1,1435;Initial Catalog=BoardDb;Persist Security Info=False;User ID=sa;Password=u7s8BkDq8lbDS3tCRPA5;TrustServerCertificate=true"
+  },
+  "Serilog": {
+    "Using": [
+      "Serilog.Sinks.Console"
+    ],
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft": "Warning",
+        "Microsoft.Hosting.Lifetime": "Information"
+      }
+    },
+    "WriteTo": [
+      {
+        "Name": "Console",
+        "Args": {
+          "formatter": "Serilog.Formatting.Json.JsonFormatter",
+          "theme": "Serilog.Sinks.SystemConsole.Themes.AnsiConsoleTheme::Literate, Serilog.Sinks.Console",
+          "outputTemplate": "[{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz}] [{Level:u3}] {MessageId} {Message:lj}{NewLine}{Exception}"
+        }
+      }
+    ],
+    "Enrich": [
+      "FromLogContext",
+      "WithEnvironmentName",
+      "WithSpan",
+      "WithExceptionDetails"
+    ]
   }
 }
