+++ src/backend/Services/Migrators/BoardDb.MigrationService/Services/BoardDbMigrationService.cs
@@ -1,4 +1,5 @@
 using Board.Infrastructure.Data;
+using Board.Infrastructure.Services;
 
 namespace BoardDb.MigrationService.Services;
